+++ craps-simulator/Interfaces/IBet.cs
@@ -5,7 +5,7 @@
         public string Name { get; }
         public float Bet { get; }
         void PlaceBet(float bet);
-        public IBetResult Result(Game game, (int die1, int die2) dice);
+        public IBetResult Result(Game game, (short die1, short die2) dice);
         public float SessionResult { get; }
     }
 }
